requirements.txt
flask==2.0.1
python-dotenv==0.19.1
requests==2.26.0
werkzeug==2.0.1
jinja2==3.0.1
itsdangerous==2.0.1
supabase==1.0.3
langchain-openai>=0.0.2
langchain-core>=0.1.0 